move enums into a dedicated package
