+++ src/main/java/edu/wpi/aquamarine_axolotls/db/RequestTable.java
@@ -1,5 +1,7 @@
 package edu.wpi.aquamarine_axolotls.db;
 
+import edu.wpi.aquamarine_axolotls.db.enums.TABLES;
+
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
