+++ src/main/java/edu/wpi/aquamarine_axolotls/db/enums/SERVICEREQUEST.java
@@ -1,4 +1,4 @@
-package edu.wpi.aquamarine_axolotls.db;
+package edu.wpi.aquamarine_axolotls.db.enums;
 
 /**
  * Enum for service request tables.
