+++ src/main/java/edu/wpi/aquamarine_axolotls/db/enums/STATUS.java
@@ -1,4 +1,4 @@
-package edu.wpi.aquamarine_axolotls.db;
+package edu.wpi.aquamarine_axolotls.db.enums;
 
 /**
  * Enum for statuses used with service requests.
