+++ src/main/java/edu/wpi/aquamarine_axolotls/views/FloralDelivery.java
@@ -4,12 +4,10 @@ import com.jfoenix.controls.*;
 import com.jfoenix.transitions.hamburger.HamburgerBasicCloseTransition;
 import edu.wpi.aquamarine_axolotls.Aapp;
 import edu.wpi.aquamarine_axolotls.db.*;
+import edu.wpi.aquamarine_axolotls.db.enums.SERVICEREQUEST;
 import javafx.collections.FXCollections;
 import javafx.fxml.FXML;
-import javafx.scene.control.ComboBox;
 import javafx.scene.layout.AnchorPane;
-
-import javafx.scene.layout.VBox;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
