+++ src/main/java/edu/wpi/aquamarine_axolotls/views/LaundryService.java
@@ -3,6 +3,7 @@ package edu.wpi.aquamarine_axolotls.views;
 import com.jfoenix.controls.*;
 import com.jfoenix.transitions.hamburger.HamburgerBasicCloseTransition;
 import edu.wpi.aquamarine_axolotls.db.*;
+import edu.wpi.aquamarine_axolotls.db.enums.SERVICEREQUEST;
 import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
