+++ src/test/java/edu/wpi/aquamarine_axolotls/db/DatabaseControllerTest4.java
@@ -1,14 +1,15 @@
 package edu.wpi.aquamarine_axolotls.db;
 
 import edu.wpi.aquamarine_axolotls.TestUtil;
-import org.apache.derby.shared.common.error.DerbySQLIntegrityConstraintViolationException;
+import edu.wpi.aquamarine_axolotls.db.enums.SERVICEREQUEST;
+import edu.wpi.aquamarine_axolotls.db.enums.STATUS;
+import edu.wpi.aquamarine_axolotls.db.enums.TABLES;
 import org.junit.jupiter.api.*;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.sql.SQLException;
 import java.util.*;
-import java.util.stream.Collectors;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static edu.wpi.aquamarine_axolotls.db.DatabaseUtil.*;
